Change from relevancy.service.ts to topics.service.ts
