+++ src/app/core/client-module/navBarDropdown.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { NavigationEnd, Router } from '@angular/router';
 import { BehaviorSubject } from 'rxjs';
 import { Topic } from '../../../entity';
-import { RelevancyService } from 'app/core/learning-object-module/relevancy/relevancy.service';
+import { TopicsService } from '../learning-object-module/topics/topics.service';
 
 
 @Injectable({
@@ -12,7 +12,7 @@ export class NavbarDropdownService {
 
     constructor(
         private router: Router,
-        private relevancyService: RelevancyService
+        private topicsService: TopicsService
     ) { }
     //mobile or desktop
     public isDesktop = new BehaviorSubject<boolean>(true);
@@ -42,7 +42,7 @@ export class NavbarDropdownService {
     public topicSelection = new BehaviorSubject<Topic>({ _id: '', name: '' });
 
     async getTopicList(): Promise<void> {
-        this.topics.next(await this.relevancyService.getTopics());
+        this.topics.next(await this.topicsService.getTopics());
     }
 
     public setTopic(topic: Topic): void {
